Allow test variables to be nested (#109)
* Allow test variables to be nested

Update the format_keys() function to change the variables dict to a Box
object. This allows the the test variables to be nested. For example:

  variables:
    users:
      foo:
        username: foo
        password: foo_pass
      bar:
        username: bar
        password: bar_pass

Then in the test spec they can be referenced as "{users.foo.username:s}"
and "{users.bar.username}".

Updated the unit tests "includes" fixture to have nested variables for
the request, and the "req" fixure for the requests test to reference the
nested variables.

* Advanced integration test includes nested variables

Split the "host" variable into 3 nested variables, and update the
advanced test spec to refernce these using dot notation.
 - service.proto (the protocol: http)
 - service.host  (the hostname: localhost)
 - service.port  (the port: 5000)