Ability to load external test spec components (stages)
Extend the test schema to permit loading external stages into any test
spec. Do this by extending the YAML parser to include two new
constructors, !load and !stage.

!load <external_file.yaml>
---------------------------
Loads test spec stages from an external YAML. The stages are just made
available to the test spec; they are not executed at this point. The
format for the file is similar to the "stages" subdocument:

  stages:
    stage_x:
      - name: Stage X
      - request: ...

!stage <stage_id>
-----------------
Extend the test spec with the stage. The YAML document is actually
extended, as if the stage were written at this spot. This means things
like anchors, saved elements, etc. are all available. Example usage
would be:

  test_name: Test ABC
  loads:
    - !load <external_file.yaml>
  stages:
    - !stage stage_x

Why do this? Anchors partially address this in the same file. This can
be a little incovenient for stages that are needed in all tests (i.e.
tests in different files). An example would be a stage that
authenticates and stores a token that will be used in subsequent stages.

The goal of this implementation is to be extensible. For example, one
may want to define a complete response or request headers externally
that can be reused across tests.

Testing
-------
New integration test added for external components (stages). It defines
an authentication stage that is then used in 2 separate tavern tests.