Support for reusable stages
Extend the includes to support definition of stages. Stages can simply
be defined in an include file just as they are defined as part of the
test spec. The schema for stages is extended to support an "id".
Included stages will define an id, which is then used to reference when
defining tests.

For example, an include file can now define resuable stages like this:
  stages:
    - id: login_as_user_foo
      request: ...
      response: ...
    - id: query_data
      request: ...
      response: ...

The "stages" part of a test spec has been extended to support stage
references. Stage reference results in the stage being copied into the
test spec as if it was defined there.

Test spec is extended to reference included stages

  includes:
    - !include common/stages.yaml
  test_name: Get data as user foo
  stages:
    - type: ref
      id: login_as_user_foo
    - name: Query data
      request: ...
      response: ...

All saved variables by referenced stages are available to subsequent
stages. So in the above example, credentials saved by
'login_as_user_foo' can be used in the rest of the test.