mqtt: fix unexpected behaviour
Currently mqtt_response with the unexpected key set to true fails if no
message is received. This is strange, was we don't expect a message, and
should only fail if a message is received.

Formatted stage:
  mqtt_publish:
    payload: 10.10.10.10
    topic: inet6/add
  mqtt_response:
    payload: !anything ''
    timeout: 5
    topic: vallumd/will
    unexpected: true

Errors:
E   tavern.util.exceptions.TestFailError: Test 'add IPv4 IP to IPv6 ipset' failed:
    - Expected '<Tavern YAML sentinel for anything>' on topic 'vallumd/will' but no such message received

Fix this by adding an extra check for the expected key when we received
no message.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>